+++ CHANGES.md
@@ -60,6 +60,9 @@
 * Change to throttle connection rate on thread initiation for adba library
 * Change default manage episodes selector to Snatched episodes if items exist else Wanted on Episode Status Manage page
 * Change snatched row colour on Episode Status Manage page to match colour used on the show details page
+* Change replace trakt with libtrakt for API v2
+* Change Trakt notification config to only handle PIN authentication with the service
+* Remove all other Trakt deprecated API V1 service features pending reconsideration
 
 [develop changelog]
 Enable Alpha Ratio again now that the secure login page over https is fixed